+++ organizations/backends/defaults.py
@@ -1,3 +1,6 @@
+"""Backend classes should provide common interface
+"""
+
 import uuid
 
 from django.conf import settings
@@ -13,21 +16,19 @@ from django.utils.translation import ugettext as _
 from ..models import get_user_model
 from ..utils import create_organization
 from ..utils import model_field_attr
-from .forms import (UserRegistrationForm,
-        OrganizationRegistrationForm)
+from .forms import UserRegistrationForm, OrganizationRegistrationForm
 from .tokens import RegistrationTokenGenerator
-
-
-# Backend classes should provide common interface
 
 
 class BaseBackend(object):
     """
     Base backend class for registering and inviting users to an organization
     """
+    org_model = None
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, org_model=None, *args, **kwargs):
         self.user_model = get_user_model()
+        self.org_model = org_model
 
     def get_urls(self):
         raise NotImplementedError
@@ -47,13 +48,36 @@ class BaseBackend(object):
         return RegistrationTokenGenerator().make_token(user)
 
     def get_username(self):
-        """Returns a UUID based 'random' and unique username"""
+        """
+        Returns a UUID-based 'random' and unique username.
+
+        This is required data for user models with a username field.
+        """
         return str(uuid.uuid4())[:model_field_attr(self.user_model, 'username', 'max_length')]
+
+    def activate_organizations(self, user):
+        """
+        Activates the related organizations for the user.
+
+        It only activates the related organizations by model type - that is, if
+        there are multiple types of organizations then only organizations in
+        the provided model class are activated.
+        """
+        try:
+            relation_name = self.org_model().user_relation_name
+        except TypeError:
+            # No org_model specified, raises a TypeError because NoneType is
+            # not callable. Thiis the most sensible default
+            relation_name = "organizations_organization"
+        organization_set = getattr(user, relation_name)
+        for org in organization_set.filter(is_active=False):
+            org.is_active = True
+            org.save()
 
     def activate_view(self, request, user_id, token):
         """
-        Activates the given User by setting `is_active` to true if the provided
-        information is verified.
+        View function that activates the given User by setting `is_active` to
+        true if the provided information is verified.
         """
         try:
             user = self.user_model.objects.get(id=user_id, is_active=False)
@@ -67,9 +91,7 @@ class BaseBackend(object):
             user = form.save()
             user.set_password(form.cleaned_data['password'])
             user.save()
-            for org in user.organization_set.filter(is_active=False):
-                org.is_active = True
-                org.save()
+            self.activate_organizations(user)
             user = authenticate(username=form.cleaned_data['username'],
                     password=form.cleaned_data['password'])
             login(request, user)
@@ -86,7 +108,8 @@ class BaseBackend(object):
         self._send_email(user, self.reminder_subject, self.reminder_body,
                 sender, **kwargs)
 
-    # TODO replace with _send_message, channel agnostic?
+    # This could be replaced with a more channel agnostic function, most likely
+    # in a custom backend.
     def _send_email(self, user, subject_template, body_template,
             sender=None, **kwargs):
         """Utility method for sending emails to new users"""