+++ organizations/managers.py
@@ -4,7 +4,11 @@ from django.db import models
 class OrgManager(models.Manager):
 
     def get_for_user(self, user):
-        return self.get_query_set().filter(users=user)
+        if hasattr(self, 'get_queryset'):
+            return self.get_queryset().filter(users=user)
+        else:
+            # Deprecated method for older versions of Django
+            return self.get_query_set().filter(users=user)
 
 
 class ActiveOrgManager(OrgManager):
@@ -13,6 +17,12 @@ class ActiveOrgManager(OrgManager):
     including only active organizations
     """
 
-    def get_query_set(self):
-        return super(ActiveOrgManager,
-                self).get_query_set().filter(is_active=True)
+    def get_queryset(self):
+        try:
+            return super(ActiveOrgManager,
+                    self).get_queryset().filter(is_active=True)
+        except AttributeError:
+            # Deprecated method for older versions of Django.
+            return super(ActiveOrgManager,
+                    self).get_query_set().filter(is_active=True)
+    get_query_set = get_queryset