+++ tests/test_models.py
@@ -49,6 +49,10 @@ class OrgModelTests(TestCase):
         self.assertTrue("{0}".format(self.foo.owner))
         self.assertTrue("{0}".format(self.foo.owner.organization_user))
 
+    def test_relation_name(self):
+        """Ensure user-related name is accessible from common attribute"""
+        self.assertEqual(self.foo.user_relation_name, "organizations_organization")
+
     def test_duplicate_members(self):
         """Ensure that a User can only have one OrganizationUser object"""
         self.assertRaises(IntegrityError, self.nirvana.add_user, self.dave)
@@ -69,6 +73,12 @@ class OrgModelTests(TestCase):
         new_guy = self.foo.add_user(self.krist)
         self.assertTrue(isinstance(new_guy, OrganizationUser))
         self.assertEqual(new_guy.organization, self.foo)
+        
+    def test_remove_user(self):
+        new_guy = self.foo.add_user(self.krist)
+        self.foo.remove_user(self.krist)
+        self.assertFalse(self.foo.users.filter(pk=self.krist.pk).exists())
+        
 
     def test_get_or_add_user(self):
         """Ensure `get_or_add_user` adds a user IFF it exists"""
@@ -142,6 +152,11 @@ class CustomModelTests(TestCase):
 
     def test_org_string(self):
         self.assertEqual(self.red_account.__str__(), "Red Account")
+
+    def test_relation_name(self):
+        """Ensure user-related name is accessible from common attribute"""
+        self.assertEqual(self.red_account.user_relation_name,
+                "test_accounts_account")
 
     def test_change_user(self):
         """Ensure custom organizations validate in owner change"""